add test for vehicle position

+++ TripUpdate/StopInfo/StopInfo.cs
@@ -15,7 +15,7 @@ namespace BusTripUpdate
             public string No { get; set; }
             public float Lat { get; set; }
             public float Lon { get; set; }
-            public float? Tm { get; set; }
+            public double? Tm { get; set; }
             public int Sid { get; set; }
             public int Seq { get; set; }
             public string Alias { get; set; }
